refactor application
