+++ App.tsx
@@ -1,67 +1,32 @@
-import React from 'react';
-import { StyleSheet, Text, View } from 'react-native';
-import { useState, useEffect } from 'react';
 import * as Permissions from 'expo-permissions';
 import * as Location from 'expo-location';
-import FireMap from './Fire-Map';
-import AirQuality from './AirQuality';
-import Navigator from './Navigator';
+
+import React, { useState, useEffect } from 'react';
+import { View } from 'react-native';
 import { NavigationContainer } from '@react-navigation/native';
-import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
+import { ApolloProvider } from '@apollo/client';
 
-import {
-  ApolloClient,
-  InMemoryCache,
-  ApolloProvider,
-  useQuery,
-  gql,
-  makeVar,
-} from '@apollo/client';
+import { TabBar } from './components/TabBar';
+import { appStyles } from './utils/style';
+import { client } from './utils/store';
+import { Coord } from './utils/types';
 
-const Tab = createBottomTabNavigator();
-
-function MyTabs() {
-  return (
-    <Tab.Navigator 
-    tabBarOptions={{
-      labelStyle: { fontSize: 18 },
-      activeTintColor: 'darkgrey',
-      style: { backgroundColor: 'white' }
-    }}>
-      <Tab.Screen name="Air" component={AirQuality} />
-      <Tab.Screen name="Fire" component={FireMap} />
-    </Tab.Navigator>
-  );
-}
-// set up Apollo Client
-const client = new ApolloClient({
-  uri: 'https://immense-escarpment-33083.herokuapp.com/graphql',
-  cache: new InMemoryCache(),
-});
-
-export const latLong: any = makeVar([]);
-
-// App Component
 export default function App() {
-  const [latitude, setLatitude] = useState(0);
-  const [longitude, setLongitude] = useState(0);
-  const [savedCoord, setSavedCoord] = useState(0);
+  const [userLocation, setUserLocation] = useState<Coord>({latitude: 34.0522, longitude: -118.2437})
 
+  // Get location permission only on initial render
   useEffect(() => {
     async function getLocationAsync() {
       try {
-        // permissions returns only for location permissions on iOS and under certain conditions, see Permissions.LOCATION
+        // permissions returns only for location permissions on iOS and under certain conditions
+        // see Permissions.LOCATION
         const { status } = await Permissions.askAsync(Permissions.LOCATION);
         if (status === 'granted') {
-          let location = await Location.getCurrentPositionAsync({});
-          setLatitude(location.coords.latitude);
-          setLongitude(location.coords.longitude);
-          setSavedCoord(Math.floor(location.coords.latitude));
-          console.log(latitude);
-          console.log(longitude);
-          console.log(savedCoord);
-          latLong([location.coords.latitude, location.coords.longitude]);
+          const location = await Location.getCurrentPositionAsync({});
+          const { latitude, longitude } = location.coords;
+          setUserLocation({ latitude, longitude })
         } else {
+          // Debugging logs
           throw new Error('Location permission not granted');
         }
       } catch(err) {
@@ -70,24 +35,15 @@ export default function App() {
     }
 
     getLocationAsync();
-  }, [savedCoord]);
+  }, []);
 
-  console.log('LatLong', latLong());
   return (
     <ApolloProvider client={client}>
-      <View style={styles.container}>
+      <View style={appStyles.container}>
         <NavigationContainer>
-          <MyTabs />
+          <TabBar userLocation={userLocation} />
         </NavigationContainer>
       </View>
     </ApolloProvider>
   );
 }
-
-// Style Sheet
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-  },
-});
-
