+++ utils/types.ts
@@ -1,0 +1,4 @@
+export interface Coord {
+  latitude: number;
+  longitude: number;
+}
